Update shop server (info model)
1. info model의 소수를 사용하는 컬럼에 값의 범위가 커서 DOUBLE로 변경

+++ Backend/Shop/models/info.js
@@ -32,18 +32,18 @@ module.exports = (sequelize, DataTypes) => {
         },
     
         grade_avg: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.DOUBLE,
             allowNULL: true,
             defaultValue: 0
         },
 
         latitude: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.DOUBLE,
             allowNULL: true
         },
         
         longitude: {
-            type: DataTypes.FLOAT,
+            type: DataTypes.DOUBLE,
             allowNULL: true
         },
